# npm-packages/@convex-dev/platform/public-deployment-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convex Public HTTP routes",
    "description": "Endpoints that require no authentication",
    "license": {
      "name": "LicenseRef-FSL-1.1-Apache-2.0",
      "identifier": "LicenseRef-FSL-1.1-Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Deployment API"
    }
  ],
  "paths": {
    "/query": {
      "get": {
        "summary": "Execute query (GET)",
        "description": "Execute a query function via GET request.",
        "operationId": "public_query_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Function path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "args",
            "in": "query",
            "description": "Function arguments as JSON string",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Response format",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UdfResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Execute query (POST)",
        "description": "Execute a query function via POST request.",
        "operationId": "public_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UdfPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UdfResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/mutation": {
      "post": {
        "summary": "Execute mutation",
        "description": "Execute a mutation function.",
        "operationId": "public_mutation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UdfResponse"
                }
              }
            }
          }
        }
      }
    },
    "/action": {
      "post": {
        "summary": "Execute action",
        "description": "Execute an action function.",
        "operationId": "public_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UdfPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UdfResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MutationPostRequest": {
        "type": "object",
        "required": [
          "path",
          "args"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "args": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/MutationPriority",
            "description": "Bounded scheduling preference. Defaults to normal and does not change\nauthorization."
          }
        }
      },
      "MutationPriority": {
        "type": "string",
        "description": "Per-call scheduling preference; it grants no additional execution authority.",
        "enum": [
          "normal",
          "high"
        ]
      },
...
      "UdfPostRequest": {
        "type": "object",
        "required": [
          "path",
          "args"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "args": {
            "type": "object"
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UdfResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "value",
              "status"
            ],
            "properties": {
              "value": {
                "$ref": "#/components/schemas/Value"
              },
              "logLines": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "success"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "errorMessage",
              "status"
            ],
            "properties": {
              "errorMessage": {
                "type": "string"
              },
              "errorData": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Value"
                  }
                ]
              },
              "logLines": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "error"
                ]
              }
            }
          }
        ]
      },
      "Value": {}
    }
  }
}